+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - fMHA: Fixed a bug in cutlass backend forward pass where the logsumexp was not correctly calculated, resulting in wrong results in the BW pass. This would happen with MQA when one sequence has a query with `length%64 == 1`
 ### Added
+- fMHA: Added `LocalAttentionFromBottomRightMask` (local)
+- fMHA: Added `LowerTriangularFromBottomRightMask` (causal)
+- fMHA: Added `LowerTriangularFromBottomRightLocalAttentionMask` (local + causal)
 ### Removed
 - Removed `xformers.triton.sum_strided`